Maybe

+++ bit_manipulation/4-clear_bit.c
@@ -5,6 +5,6 @@ int clear_bit(unsigned long int *n, unsigned int index)
 	unsigned long int mask;
 
 	mask = ~(1UL << index)
-	*n = *n & mask;
+	*n = (unsigned long int)(*n & mask);
 	return 1;
 }
